+++ abi/abi_decode.go
@@ -3,17 +3,15 @@ package abi
 import (
 	"errors"
 	"fmt"
-	"reflect"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
 )
 
+const tupleArrayType = "tuple[]"
+
 // DecodeABIEncodedStructArray is a generic helper that decodes ABI-encoded tuple array
 // It handles the ABI unpacking and type conversion boilerplate
-func DecodeABIEncodedStructArray[T any](
-	encodedBytes []byte,
-	converter func(any) (T, error),
-) ([]T, error) {
+func DecodeABIEncodedStructArray[T any](encodedBytes []byte) ([]T, error) {
 	if len(encodedBytes) == 0 {
 		return nil, errors.New("encoded bytes are empty")
 	}
@@ -24,7 +22,7 @@ func DecodeABIEncodedStructArray[T any](
 		return nil, fmt.Errorf("failed to build ABI fields: %w", err)
 	}
 
-	arrayType, err := abi.NewType("tuple[]", "", abiFields)
+	arrayType, err := abi.NewType(tupleArrayType, "", abiFields)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create array type: %w", err)
 	}
@@ -40,22 +38,6 @@ func DecodeABIEncodedStructArray[T any](
 		return nil, errors.New("unpacked data is empty")
 	}
 
-	// The unpacked[0] contains the slice, but we need to extract it via reflection
-	// since the ABI library returns anonymous structs
-	val := reflect.ValueOf(unpacked[0])
-	if val.Kind() != reflect.Slice {
-		return nil, fmt.Errorf("expected slice, got %v", val.Kind())
-	}
-
-	result := make([]T, val.Len())
-	for i := 0; i < val.Len(); i++ {
-		item := val.Index(i).Interface()
-		converted, err := converter(item)
-		if err != nil {
-			return nil, fmt.Errorf("failed to convert item %d: %w", i, err)
-		}
-		result[i] = converted
-	}
-
-	return result, nil
+	decodedData := *abi.ConvertType(unpacked[0], new([]T)).(*[]T)
+	return decodedData, nil
 }
