+++ abi/abi_decode_test.go
@@ -1,7 +1,6 @@
 package abi
 
 import (
-	"errors"
 	"math/big"
 	"testing"
 
@@ -11,9 +10,9 @@ import (
 
 func TestDecodeABIEncodedStructArray(t *testing.T) {
 	type TestStruct struct {
-		Field1 uint8          `abiarg:"field1"`
-		Field2 uint32         `abiarg:"field2"`
-		Field3 common.Address `abiarg:"field3"`
+		Field1 uint8          `abi:"field1"`
+		Field2 uint32         `abi:"field2"`
+		Field3 common.Address `abi:"field3"`
 	}
 
 	// Create test data
@@ -35,66 +34,25 @@ func TestDecodeABIEncodedStructArray(t *testing.T) {
 	require.NoError(t, err)
 	require.NotEmpty(t, encodedBytes)
 
-	// Decode with converter
-	converter := func(item any) (TestStruct, error) {
-		// The ABI library returns anonymous structs, we need to extract fields
-		// In real usage, you'd use reflection or type assertions
-		return TestStruct{
-			Field1: 1, // Placeholder for test
-			Field2: 100,
-			Field3: common.HexToAddress("0x1111111111111111111111111111111111111111"),
-		}, nil
-	}
-
-	decoded, err := DecodeABIEncodedStructArray(encodedBytes, converter)
+	decoded, err := DecodeABIEncodedStructArray[TestStruct](encodedBytes)
 	require.NoError(t, err)
 	require.Len(t, decoded, 2)
 }
 
 func TestDecodeABIEncodedStructArray_EmptyBytes(t *testing.T) {
 	type TestStruct struct {
-		Field1 uint8 `abiarg:"field1"`
+		Field1 uint8 `abi:"field1"`
 	}
 
-	converter := func(item any) (TestStruct, error) {
-		return TestStruct{}, nil
-	}
-
-	_, err := DecodeABIEncodedStructArray([]byte{}, converter)
+	_, err := DecodeABIEncodedStructArray[TestStruct]([]byte{})
 	require.Error(t, err)
 	require.Contains(t, err.Error(), "encoded bytes are empty")
 }
 
-func TestDecodeABIEncodedStructArray_ConverterError(t *testing.T) {
-	type TestStruct struct {
-		Field1 uint8  `abiarg:"field1"`
-		Field2 uint32 `abiarg:"field2"`
-	}
-
-	// Create test data
-	items := []TestStruct{
-		{Field1: 1, Field2: 100},
-		{Field1: 2, Field2: 200},
-	}
-
-	encodedBytes, err := EncodeABIStructArray(items)
-	require.NoError(t, err)
-
-	// Converter that always fails
-	converter := func(item any) (TestStruct, error) {
-		return TestStruct{}, errors.New("converter failed")
-	}
-
-	_, err = DecodeABIEncodedStructArray(encodedBytes, converter)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "failed to convert item 0")
-	require.Contains(t, err.Error(), "converter failed")
-}
-
 func TestDecodeABIEncodedStructArray_WithBigInt(t *testing.T) {
 	type TestStruct struct {
-		Amount *big.Int `abiarg:"amount,uint256"`
-		Value  uint32   `abiarg:"value"`
+		Amount *big.Int `abi:"amount"`
+		Value  uint32   `abi:"value"`
 	}
 
 	// Create test data
@@ -107,20 +65,14 @@ func TestDecodeABIEncodedStructArray_WithBigInt(t *testing.T) {
 	encodedBytes, err := EncodeABIStructArray(items)
 	require.NoError(t, err)
 
-	// Converter that extracts fields
-	converter := func(item any) (TestStruct, error) {
-		// In real usage, you'd use reflection to extract the fields
-		return TestStruct{Amount: big.NewInt(1000), Value: 1}, nil
-	}
-
-	decoded, err := DecodeABIEncodedStructArray(encodedBytes, converter)
+	decoded, err := DecodeABIEncodedStructArray[TestStruct](encodedBytes)
 	require.NoError(t, err)
 	require.Len(t, decoded, 3)
 }
 
 func TestDecodeABIEncodedStructArray_EmptyArray(t *testing.T) {
 	type TestStruct struct {
-		Field1 uint8 `abiarg:"field1"`
+		Field1 uint8 `abi:"field1"`
 	}
 
 	// Encode empty array
@@ -128,41 +80,33 @@ func TestDecodeABIEncodedStructArray_EmptyArray(t *testing.T) {
 	encodedBytes, err := EncodeABIStructArray(items)
 	require.NoError(t, err)
 
-	converter := func(item any) (TestStruct, error) {
-		return TestStruct{}, nil
-	}
-
-	decoded, err := DecodeABIEncodedStructArray(encodedBytes, converter)
+	decoded, err := DecodeABIEncodedStructArray[TestStruct](encodedBytes)
 	require.NoError(t, err)
 	require.Len(t, decoded, 0)
 }
 
 func TestDecodeABIEncodedStructArray_InvalidABIData(t *testing.T) {
 	type TestStruct struct {
-		Field1 uint8 `abiarg:"field1"`
-	}
-
-	converter := func(item any) (TestStruct, error) {
-		return TestStruct{}, nil
+		Field1 uint8 `abi:"field1"`
 	}
 
 	// Invalid ABI encoded data
 	invalidData := []byte{0x01, 0x02, 0x03}
 
-	_, err := DecodeABIEncodedStructArray(invalidData, converter)
+	_, err := DecodeABIEncodedStructArray[TestStruct](invalidData)
 	require.Error(t, err)
 	require.Contains(t, err.Error(), "failed to unpack data")
 }
 
 func TestDecodeABIEncodedStructArray_ComplexStruct(t *testing.T) {
 	type ComplexStruct struct {
-		LeafType           uint8          `abiarg:"leafType"`
-		OriginNetwork      uint32         `abiarg:"originNetwork"`
-		OriginAddress      common.Address `abiarg:"originAddress"`
-		DestinationNetwork uint32         `abiarg:"destinationNetwork"`
-		DestinationAddress common.Address `abiarg:"destinationAddress"`
-		Amount             *big.Int       `abiarg:"amount,uint256"`
-		Metadata           []byte         `abiarg:"metadata"`
+		LeafType           uint8          `abi:"leafType"`
+		OriginNetwork      uint32         `abi:"originNetwork"`
+		OriginAddress      common.Address `abi:"originAddress"`
+		DestinationNetwork uint32         `abi:"destinationNetwork"`
+		DestinationAddress common.Address `abi:"destinationAddress"`
+		Amount             *big.Int       `abi:"amount"`
+		Metadata           []byte         `abi:"metadata"`
 	}
 
 	// Create test data
@@ -190,16 +134,7 @@ func TestDecodeABIEncodedStructArray_ComplexStruct(t *testing.T) {
 	encodedBytes, err := EncodeABIStructArray(items)
 	require.NoError(t, err)
 
-	converter := func(item any) (ComplexStruct, error) {
-		// Placeholder converter for test
-		return ComplexStruct{
-			LeafType:      1,
-			OriginNetwork: 1,
-			Amount:        big.NewInt(1000),
-		}, nil
-	}
-
-	decoded, err := DecodeABIEncodedStructArray(encodedBytes, converter)
+	decoded, err := DecodeABIEncodedStructArray[ComplexStruct](encodedBytes)
 	require.NoError(t, err)
 	require.Len(t, decoded, 2)
 }
@@ -210,13 +145,9 @@ func TestDecodeABIEncodedStructArray_NoABITags(t *testing.T) {
 		Field2 uint32
 	}
 
-	converter := func(item any) (BadStruct, error) {
-		return BadStruct{}, nil
-	}
-
-	// Try to decode with a struct that has no abiarg tags
+	// Try to decode with a struct that has no abi tags
 	// BuildABIFields will succeed but return empty fields, which will cause unpack to fail
-	_, err := DecodeABIEncodedStructArray([]byte{0x01}, converter)
+	_, err := DecodeABIEncodedStructArray[BadStruct]([]byte{0x01})
 	require.Error(t, err)
 	// The error will be from unpacking due to insufficient data or empty ABI fields
 	require.Contains(t, err.Error(), "failed to")
@@ -224,7 +155,7 @@ func TestDecodeABIEncodedStructArray_NoABITags(t *testing.T) {
 
 func TestDecodeABIEncodedStructArray_SingleItem(t *testing.T) {
 	type TestStruct struct {
-		Value uint64 `abiarg:"value"`
+		Value uint64 `abi:"value"`
 	}
 
 	// Create single item array
@@ -235,42 +166,8 @@ func TestDecodeABIEncodedStructArray_SingleItem(t *testing.T) {
 	encodedBytes, err := EncodeABIStructArray(items)
 	require.NoError(t, err)
 
-	converter := func(item any) (TestStruct, error) {
-		return TestStruct{Value: 12345}, nil
-	}
-
-	decoded, err := DecodeABIEncodedStructArray(encodedBytes, converter)
+	decoded, err := DecodeABIEncodedStructArray[TestStruct](encodedBytes)
 	require.NoError(t, err)
 	require.Len(t, decoded, 1)
 	require.Equal(t, uint64(12345), decoded[0].Value)
-}
-
-func TestDecodeABIEncodedStructArray_ConverterPartialFailure(t *testing.T) {
-	type TestStruct struct {
-		Field1 uint8  `abiarg:"field1"`
-		Field2 uint32 `abiarg:"field2"`
-	}
-
-	items := []TestStruct{
-		{Field1: 1, Field2: 100},
-		{Field1: 2, Field2: 200},
-		{Field1: 3, Field2: 300},
-	}
-
-	encodedBytes, err := EncodeABIStructArray(items)
-	require.NoError(t, err)
-
-	callCount := 0
-	converter := func(item any) (TestStruct, error) {
-		callCount++
-		if callCount == 2 {
-			return TestStruct{}, errors.New("failed on item 2")
-		}
-		return TestStruct{Field1: 1, Field2: 100}, nil
-	}
-
-	_, err = DecodeABIEncodedStructArray(encodedBytes, converter)
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "failed to convert item 1")
-	require.Contains(t, err.Error(), "failed on item 2")
 }
